FillToolHelperクラスを追加し、FillStartメソッドを実装して塗りつぶし処理を整理
